trocando unsigned int por size_t outras coisas

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: rde-lima <rde-lima@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/16 21:59:40 by rde-lima          #+#    #+#             */
-/*   Updated: 2021/05/18 21:41:33 by rde-lima         ###   ########.fr       */
+/*   Updated: 2021/05/30 03:56:34 by rde-lima         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while (str[length] != '\0')
