+++ ft_strncmp.c
@@ -6,11 +6,13 @@
 /*   By: rde-lima <rde-lima@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/23 14:28:32 by rde-lima          #+#    #+#             */
-/*   Updated: 2021/05/23 14:30:45 by rde-lima         ###   ########.fr       */
+/*   Updated: 2021/05/30 03:57:21 by rde-lima         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	if (n == 0)
 		return (n);
